+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   aren't initialized before use.
 - `DateFormatSettingsRule` analysis rule, which flags places where the default `TFormatSettings` is
   implicitly used.
+- `ForbiddenFieldRule` analysis rule template, which flags fields from a specified list of names on
+  a type with a specified fully-qualified name.
 
 ### Changed